[frontend] Disable song finder
I've gotten tired of the backend taking up a bunch of precious RAM
and therefore I'm removing song finding until I can find time to
rewrite it in a lower level language that isn't as memory hungry
(and depends on so many modules). Meanwhile the frontend functions
fine even without the discovery backend with a little tweaking.